Fix boundary

+++ weighted_imputation/algorithms/nodes.py
@@ -89,7 +89,7 @@ def boundary(graph: Graph, nodes: List[str]) -> List[str]:
     _nodes = graph.nodes()
     if not set(nodes).issubset(set(_nodes)):
         raise Exception('node not in graph.')
-    indices = np.array([_nodes.index(node) for node in nodes])
+    indices = np.array([_nodes.index(node) for node in nodes], dtype=int)
     adjacency_matrix = graph.adjacency_matrix(copy=False)
     boundary = _boundary(indices, adjacency_matrix)
     boundary = [_nodes[bound] for bound in boundary]
